modified install_docker.sh

+++ linux_install_scripts/install_docker.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 
+if [ "$#" -lt 1 ];then
+	echo "Usage: sudo $0 \$USER"
+    exit 1
+fi
+
+echo "installing docker"
+
 #remove older version
 sudo apt-get remove -y docker docker-engine docker.io
 
-sudo apt-get update
+sudo apt-get update >/dev/null
 
 #offical method
 #sudo apt-get install -y docker-ce
@@ -11,10 +18,13 @@ sudo apt-get update
 #other method
 wget -q -O-  https://get.docker.com/ |sh
 
-sudo usermod -aG docker $USER
+
+sudo usermod -aG docker $1
+
+# fix error: Temporary failure resolving 'security.ubuntu.com'
+if [ "$2" = "-uwo" ]; then
+   echo {"dns": ["129.100.254.134", "129.100.254.135"]} >/etc/docker/daemon.json
+fi
 
 # test
 sudo docker run hello-world
-
-#echo 'Must:log out and then log in'
-#echo 'use "docker run hello-world" to test it.'
